refactor: Phase 1-2 중복 코드 통합 및 미사용 파일 삭제
- tech-stack-utils.ts 생성 (normalizeTechKey/findTechBadge 6곳→1곳 통합)
- RecommendDeveloperCard, BookmarkDeveloperCard, ProfileBase, TechStackChips 중복 제거
- 미사용 파일 삭제: ProjectMd, RecommendProjectBase, MainProjectLg, useAuth(mock)
- header.tsx에서 삭제된 useAuth import 정리
- docs: 파일구조 개편 계획서 추가
