+++ src/shared/components/common/RecommendDeveloperCard.tsx
@@ -4,16 +4,7 @@ import AvatarIcon from '@assets/icons/avatar.svg?react';
 import { cn } from '@libs/cn';
 import type { BadgeTone } from '@t/badgeTone';
 import { useThemeStore } from '@store/theme';
-import {
-  BACKEND_DATABASE,
-  BACKEND_FRAMEWORK,
-  BACKEND_LANGUAGE,
-  FRONTEND_LANGUAGE_FRAMEWORK,
-  FRONTEND_MOBILE,
-  INFRA_CLOUD,
-  INFRA_CONTAINER,
-  type TechStackChip,
-} from '@constants/position-tech-stack';
+import { normalizeTechKey, findTechBadge, SKIP_TECH_NAMES } from '@libs/tech-stack-utils';
 
 export type RecommendDeveloperCardTech = { id: string; name: string; icon?: React.ReactNode };
 
@@ -48,15 +39,7 @@ export type RecommendDeveloperCardProps = {
   onNavigateToDeveloper?: (nickname: string) => void;
 };
 
-const SKIP_TECH_NAMES = new Set([
-  'backend',
-  'frontend',
-  'infra',
-  '백엔드',
-  '프론트엔드',
-  '프런트엔드',
-  '인프라',
-]);
+
 
 function RecommendDeveloperCard({
   role: _role,
@@ -79,45 +62,6 @@ function RecommendDeveloperCard({
   onNavigateToDeveloper,
 }: RecommendDeveloperCardProps) {
   const { theme } = useThemeStore();
-
-  const normalizeTechKey = (v: unknown): string => {
-    const s = typeof v === 'string' ? v : v != null ? String(v) : '';
-    return s
-      .trim()
-      .toLowerCase()
-      .replace(/\s/g, '')
-      .replace(/\./g, '')
-      .replace(/-/g, '')
-      .replace(/_/g, '');
-  };
-
-  const ALL_TECH_STACK_BADGES: Array<Extract<TechStackChip, { off: string; on: string }>> = [
-    ...FRONTEND_LANGUAGE_FRAMEWORK,
-    ...FRONTEND_MOBILE,
-    ...BACKEND_LANGUAGE,
-    ...BACKEND_FRAMEWORK,
-    ...BACKEND_DATABASE,
-    ...INFRA_CLOUD,
-    ...INFRA_CONTAINER,
-  ].filter((b): b is Extract<TechStackChip, { off: string; on: string }> => 'off' in b && 'on' in b);
-
-  const TECH_BADGE_BY_NAME = new Map(
-    ALL_TECH_STACK_BADGES.flatMap((b) => [
-      [normalizeTechKey(b.key), b],
-      [normalizeTechKey(b.label), b],
-    ]),
-  );
-
-  const findBadge = (name: unknown) => {
-    const normalized = normalizeTechKey(name);
-    const alias = normalized
-      .replace(/^spring$/g, 'springboot')
-      .replace(/typescript/g, 'typescript')
-      .replace(/nextjs/g, 'nextjs')
-      .replace(/nodejs/g, 'nodejs')
-      .replace(/reactnative/g, 'reactnative');
-    return TECH_BADGE_BY_NAME.get(alias) ?? TECH_BADGE_BY_NAME.get(normalized) ?? null;
-  };
 
   const filteredTechStack =
     techStack?.filter((t) => !SKIP_TECH_NAMES.has(normalizeTechKey(t.name))) ?? [];
@@ -190,7 +134,7 @@ function RecommendDeveloperCard({
       <div className="absolute left-[780px] top-1/2 w-[300px] -translate-y-1/2">
         <div className="flex min-h-[72px] flex-wrap items-center gap-[4px]">
         {techChips.map((t) => {
-          const badge = findBadge(t.name);
+          const badge = findTechBadge(t.name);
           if (badge) {
             const offSrc = theme === 'dark' ? (badge.offDark ?? badge.off) : badge.off;
             return (
