--- src/shared/components/common/RecommendProjectBase.tsx
@@ -1,159 +1,0 @@
-import BookmarkButton from '@components/common/BookmarkButton';
-import PersonIcon from '@assets/icons/person.svg?react';
-import { badgeToneToClass } from 'src/shared/types/badgeTone';
-import type { RecommendProjectCardBaseProps } from 'src/shared/types/recommendProjectCard.types';
-
-/**
- * Recommend 전용 ProjectBase 복제본.
- * - 현재는 `ProjectBase`와 동일 구현(파일 분리 목적)
- * - 이후 Recommend 요구사항에 맞춰 독립적으로 수정 가능
- */
-export default function RecommendProjectBase(props: RecommendProjectCardBaseProps) {
-  const {
-    categoryLabel,
-    deadlineLabel,
-    thumbnailUrl,
-    thumbnailAlt,
-    title,
-    location,
-    period,
-    mode,
-    roles,
-    dueLabel,
-    bookmarked = false,
-    onBookmarkChange,
-    onClick,
-    render,
-  } = props;
-
-  const metaText = [location, period, mode].filter(Boolean).join(' · ');
-
-  const Thumbnail = (
-    <div className="shrink-0">
-      {thumbnailUrl ? (
-        <img
-          src={thumbnailUrl}
-          alt={thumbnailAlt ?? title}
-          className="size-[124px] rounded-xl bg-card-section-bg object-cover"
-          loading="lazy"
-        />
-      ) : (
-        <div className="size-[124px] rounded-xl bg-card-profile-reason-bg" />
-      )}
-    </div>
-  );
-
-  const HeaderBadges =
-    categoryLabel || deadlineLabel ? (
-      <div className="flex flex-wrap items-center gap-4">
-        {categoryLabel && (
-          <span className="inline-flex rounded-lg bg-badge-bg-gray px-3 py-1 font-medium text-badge-text-gray text-lg">
-            {categoryLabel}
-          </span>
-        )}
-        {deadlineLabel && (
-          <span className="inline-flex rounded-lg bg-badge-bg-gray px-3 py-1 font-medium text-badge-text-gray text-lg">
-            {deadlineLabel}
-          </span>
-        )}
-      </div>
-    ) : null;
-
-  const Title = (
-    <h3 className="line-clamp-2 pl-1 font-semibold text-[15px] text-card-title leading-snug">
-      {title}
-    </h3>
-  );
-
-  const Meta = metaText ? (
-    <div className="truncate pl-1 text-badge-text-gray text-lg">{metaText}</div>
-  ) : null;
-
-  // 배지 / 인원 / 구분선 / 아이콘
-  const RolesLg = roles?.length ? (
-    <div className="flex flex-col gap-y-5">
-      {roles.slice(0, 3).map((r) => (
-        <div
-          key={r.key}
-          className="grid grid-cols-[60px_auto_8px_1fr] items-center gap-x-4 text-card-muted"
-        >
-          <span
-            className={`inline-flex w-fit items-center whitespace-nowrap rounded-lg px-3 py-1 font-semibold text-base ${badgeToneToClass[r.tone]}`}
-          >
-            {r.label}
-          </span>
-
-          <div className="flex items-center gap-2">
-            <PersonIcon className="h-5 w-5 text-card-muted" />
-            <span className="whitespace-nowrap font-bold text-lg">
-              {r.current}/{r.total}
-            </span>
-          </div>
-
-          <span className="justify-self-center text-card-muted/50">
-            {r.techStack?.length ? '|' : ''}
-          </span>
-
-          <div className="flex items-center gap-2">
-            {r.techStack?.slice(0, 5).map((t) => (
-              <span key={t.id} className="inline-flex h-6 w-6 items-center justify-center">
-                {t.icon}
-              </span>
-            ))}
-          </div>
-        </div>
-      ))}
-    </div>
-  ) : null;
-
-  const RolesMd = roles?.length ? (
-    <div className="grid grid-cols-2 gap-x-10 gap-y-6">
-      {roles.slice(0, 2).map((r) => (
-        <div key={r.key} className="flex flex-col gap-3">
-          <span
-            className={`inline-flex w-fit items-center whitespace-nowrap rounded-lg px-3 py-1 font-semibold text-base ${badgeToneToClass[r.tone]}`}
-          >
-            {r.label}
-          </span>
-
-          <div className="flex items-center gap-4 text-card-muted">
-            <div className="flex items-center gap-2">
-              <PersonIcon className="h-5 w-5 text-card-muted" />
-              <span className="font-bold text-lg">
-                {r.current}/{r.total}
-              </span>
-            </div>
-
-            {r.techStack?.length ? (
-              <div className="flex items-center gap-2">
-                {r.techStack.slice(0, 5).map((t) => (
-                  <span key={t.id} className="inline-flex h-6 w-6 items-center justify-center">
-                    {t.icon}
-                  </span>
-                ))}
-              </div>
-            ) : null}
-          </div>
-        </div>
-      ))}
-    </div>
-  ) : null;
-
-  const Due = dueLabel ? <p className="w-[65px] text-badge-text-gray text-lg">{dueLabel}</p> : null;
-
-  const Bookmark = (
-    <BookmarkButton
-      bookmarked={bookmarked}
-      onBookmarkChange={onBookmarkChange}
-      stopPropagation={!!onClick}
-      className="shrink-0"
-      iconClassName="h-10 w-10"
-    />
-  );
-
-  return (
-    <>
-      {render({ metaText, Thumbnail, HeaderBadges, Title, Meta, RolesLg, RolesMd, Due, Bookmark })}
-    </>
-  );
-}
