+++ src/shared/components/profileDetail/TechStackChips.tsx
@@ -1,15 +1,6 @@
 import { useThemeStore } from '@store/theme';
 import {cn} from '@libs/cn';
-import {
-  BACKEND_DATABASE,
-  BACKEND_FRAMEWORK,
-  BACKEND_LANGUAGE,
-  FRONTEND_LANGUAGE_FRAMEWORK,
-  FRONTEND_MOBILE,
-  INFRA_CLOUD,
-  INFRA_CONTAINER,
-  type TechStackChip,
-} from '@constants/position-tech-stack';
+import { findTechBadge } from '@libs/tech-stack-utils';
 import type { TechStackSource } from '@t/profileCard.types';
 import CheckFillIcon from "@assets/icons/check-fill.svg?react";
 
@@ -43,47 +34,10 @@ const TechStackChips = ({ techStack = [] }: TechStackChipsProps) => {
   const { theme } = useThemeStore();
   const displayItems = chipInputsToDisplayItems(techStack);
 
-  const normalizeTechKey = (value: string) =>
-    value
-      .trim()
-      .toLowerCase()
-      .replaceAll(' ', '')
-      .replaceAll('.', '')
-      .replaceAll('-', '')
-      .replaceAll('_', '');
-
-  const ALL_TECH_STACK_BADGES: Array<
-    Extract<TechStackChip, { off: string; on: string }>
-  > = [
-    ...FRONTEND_LANGUAGE_FRAMEWORK,
-    ...FRONTEND_MOBILE,
-    ...BACKEND_LANGUAGE,
-    ...BACKEND_FRAMEWORK,
-    ...BACKEND_DATABASE,
-    ...INFRA_CLOUD,
-    ...INFRA_CONTAINER,
-  ].filter(
-    (badge): badge is Extract<TechStackChip, { off: string; on: string }> =>
-      'off' in badge && 'on' in badge,
-  );
-
-  const TECH_BADGE_BY_NAME = new Map<
-    string,
-    Extract<TechStackChip, { off: string; on: string }>
-  >(
-    ALL_TECH_STACK_BADGES.flatMap((badge) => [
-      [normalizeTechKey(badge.key), badge],
-      [normalizeTechKey(badge.label), badge],
-    ]),
-  );
-
-  const findBadge = (name: string) =>
-    TECH_BADGE_BY_NAME.get(normalizeTechKey(name)) ?? null;
-
   return (
     <div className="flex flex-wrap items-center gap-[4px]">
       {displayItems.map(({ name, source }, index) => {
-        const badge = findBadge(name);
+        const badge = findTechBadge(name);
 
         if (badge) {
           const offSrc =
